:recycle: Convert useSessionStorage to TypeScript.
